Add files via upload
The Online Cab Booking Application is a full-stack web application designed to provide users with an easy and efficient way to book cabs online. This system offers functionality for multiple user roles such as administrators, customers, and drivers, making it a robust solution for managing cab bookings. It utilizes Java, Spring Boot, and various other technologies to ensure high performance and scalability.